CNN-Transfer-Learning/vgg/requirements.txt
tensorflow
tensorflow-gpu
pandas
Pillow
keras
flask